+++ services/data_plane_manager/src/main/java/com/futurewei/alcor/dataplane/utils/GoalStateManager.java
@@ -91,7 +91,7 @@ public class GoalStateManager {
     com.futurewei.alcor.web.entity.vpc.VpcEntity[] vpcArr =
         networkConfiguration.getVpcs().toArray(new com.futurewei.alcor.web.entity.vpc.VpcEntity[0]);
 
-    // TODO need to refactor subnet and vpc part when logic is
+    // TODO need to v2 subnet and vpc part when logic is
     //  clear and integration done
     Map<String, List<InternalPortEntity>> mapGroupedByHostIp = new HashMap();
     Map<String, InternalSubnetEntity> subnetMap = new HashMap<>();
